Refactor the front-end queries so that each query hook is its own file
Same motivation as the back-end refactor - makes finding things a lot easier.
