+++ apps/front-end/src/resources/Users/pages/UserProfile/index.tsx
@@ -12,7 +12,7 @@ import DropdownMenuIconButton from "src/components/DropdownIconButton";
 import QueryBoundaryItemWrapper from "src/groups/QueryBoundary/QueryBoundaryWrapper";
 import AboutUser from "src/resources/Users/pages/UserProfile/AboutUser";
 import UserBlogs from "src/resources/Users/pages/UserProfile/UserBlogs";
-import { useUserQuery } from "src/resources/Users/queries";
+import useUserQuery from "src/resources/Users/queries/useUserQuery";
 
 interface UserProfileProps {
   userId: string;
